Double Stack working well

+++ Project/Split into files/image.h
@@ -3,14 +3,14 @@
 class image_sfml
 {
 private:
-    long double x_shift;
-    long double y_shift;
+    int x_shift;
+    int y_shift;
     int resolution;
     float zoom; 
     sf::VertexArray vertexarray;
 
 public:
     image_sfml();
-    void save_image_parameters(long double x_shift, long double y_shift, int resolution,float zoom,sf::VertexArray vertexarray);
-    void retrieve_image_parameters(long double &x_shift, long double &y_shift, int &resolution,float &zoom,sf::VertexArray &vertexarray);
+    void save_image_parameters(int x_shift, int y_shift, int resolution,float zoom,sf::VertexArray vertexarray);
+    void retrieve_image_parameters(int &x_shift, int &y_shift, int &resolution,float &zoom,sf::VertexArray &vertexarray);
 };
